[#171994544] Refactoring biometrics function
Refactoring biometrics function
